+++ Program/CCode/include/energy1.h
@@ -3,7 +3,7 @@
 #ifndef ENERGY_1_H_INCLUDE
 #define ENERGY_1_H_INCLUDE
 
-float energy1Wat(float tWat, struct parameters params);
-float energy1PCM(float tPCM, struct parameters params);
+double energy1Wat(double tWat, struct parameters params);
+double energy1PCM(double tPCM, struct parameters params);
 
 #endif // ENERGY_1_INCLUDE
