+++ Program/CCode/include/energy2.h
@@ -3,8 +3,8 @@
 #ifndef ENERGY_2_H_INCLUDE
 #define ENERGY_2_H_INCLUDE
 
-float energy2Wat(float tWat, struct parameters params);
-float energy2PCM(float latentHeat, struct parameters params);
+double energy2Wat(double tWat, struct parameters params);
+double energy2PCM(double latentHeat, struct parameters params);
 
 #endif // ENERGY_2_INCLUDE
